# apispec.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "NodeJS Restify REST API",
        "description": "API spec for NodeJS Restify REST API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081"
        }
    ],
    "paths": {
        "/auth": {
            "post": {
                "tags": [
                    "Auth API"
                ],
                "summary": "Login API",
                "description": "Login API",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "username",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "iat": {
                                            "type": "number"
                                        },
                                        "exp": {
                                            "type": "number"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/questions": {
            "get": {
                "tags": [
                    "Question API"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List All Questions",
                "description": "List All Questions Data",
                "responses": {
                    "200": {
                        "description": "Success Get All Questions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Question"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Question API"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Create New Question",
                "description": "Create New Question",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrUpdateQuestion"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Create Question",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Question"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                }
            }
        },
        "/questions/{questionId}": {
            "get": {
                "tags": [
                    "Question API"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Get Question by Id",
                "description": "Get Question by Id",
                "parameters": [
                    {
                        "name": "questionId",
                        "in": "path",
                        "description": "Question Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Find Question",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Question"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "Question API"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Update Question by Id",
                "description": "Update Question by Id",
                "parameters": [
                    {
                        "name": "questionId",
                        "in": "path",
                        "description": "Question Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrUpdateQuestion"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Update Question",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Question"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Question API"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Delete Question by Id",
                "description": "Delete Question by Id",
                "parameters": [
                    {
                        "name": "questionId",
                        "in": "path",
                        "description": "Question Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Delete Question",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                }
            }
        },
        "/answers": {
            "get": {
                "tags": [
                    "Answer API"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List All Answers",
                "description": "List All Answers Data",
                "responses": {
                    "200": {
                        "description": "Success Get All Answers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Answer"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Answer API"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Create New Answer",
                "description": "Create New Answer",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAnswer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Create Answer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Answer"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                }
            }
        },
        "/answers/{answerId}": {
            "get": {
                "tags": [
                    "Answer API"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Get Answer by Id",
                "description": "Get Answer by Id",
                "parameters": [
                    {
                        "name": "answerId",
                        "in": "path",
                        "description": "Answer Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Find Answer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Answer"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "Answer API"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Update Answer by Id",
                "description": "Update Answer by Id",
                "parameters": [
                    {
                        "name": "answerId",
                        "in": "path",
                        "description": "Answer Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAnswer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Update Answer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Answer"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Answer API"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Delete Answer by Id",
                "description": "Delete Answer by Id",
                "parameters": [
                    {
                        "name": "answerId",
                        "in": "path",
                        "description": "Answer Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Delete Answer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Question": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "question": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    }
                }
            },
            "CreateOrUpdateQuestion": {
                "type": "object",
                "properties": {
                    "question": {
                        "type": "string"
                    }
                },
                "required": [
                    "question"
                ]
            },
            "Answer": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "question_id": {
                        "type": "number"
                    },
                    "answer": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    }
                }
            },
            "CreateAnswer": {
                "type": "object",
                "properties": {
                    "question_id": {
                        "type": "number"
                    },
                    "answer": {
                        "type": "string"
                    }
                },
                "required": [
                    "question_id",
                    "answer"
                ]
            },
            "UpdateAnswer": {
                "type": "object",
                "properties": {
                    "answer": {
                        "type": "string"
                    }
                },
                "required": [
                    "answer"
                ]
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Bearer",
                "description": "Authentication JWT for API"
            }
        },
        "responses": {
            "NotFound": {
                "description": "Error: Record not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "number"
                                },
                                "status": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "Error: Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}